Move chance commands to baseCmd, delete brit
